stats: save uptime to know when dump had happened
We want to use a simple fact: If we have an alive process in dumpable
pstree and a starttime of that process is less than pre-dump's
timestamp, then these exact process existed at the time of these
pre-dump and the process' memory was dumped in images. Thus we save
uptime while in freezed state else these won't work.