mount: add external_slavery detection
If an external mount before checkpointing had an external master_id, it
should also have this after restoring and in contrary if it had internal
master_id it should be internal after restoring. Now we always assume
that if external mount has any master_id than these master_id is
external but it's not true.

For instance if we give an external private mount to a CT, the CT user
can make these mount shared, bind it somethere, and make these mount
slave. These way external mount can have internal master.

We already detect real state of internal/external master in
resolve_shared_mounts, so save it and use it in do_bind_mount.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>