mount: add mntinfo_add_list_before helper for adding to mntinfo list
Use this helper everywhere instead of manually adding mounts to the head
of the list, this way it is much easier to track all places where we do
add to mntinfo list.

Signed-off-by: Alexander Mikhalitsyn (Virtuozzo) <alexander@mihalicyn.com>

Cherry-picked from Virtuozzo criu:
https://src.openvz.org/projects/OVZ/repos/criu/commits/7bca9397b

Changes: reword commit message as orignaly it was a fixup but now it is
a preparational commit, don't add root_yard_mp to the list because it is
not fully initialized (e.g. lookup_nsid_by_mnt_id would be broken by it).

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>