mount: fix failed assertion in mnt_is_external
We have caught a bug with triggered BUG_ON in mnt_is_external
when c/r container with binfmt_misc mount.
The problem is that root_yard_mp (mountpoint) is not added to
mntinfo linked list. It caused the situation when __search_bindmounts
function is not processed all mountpoints and follows to the
assertion failure in mnt_is_external/__mnt_get_external.

https://jira.sw.ru/browse/PSBM-102791

Signed-off-by: Alexander Mikhalitsyn (Virtuozzo) <alexander@mihalicyn.com>