zdtm: enable mounts-v2 on restore with --mounts-v2 option
Temporary disable bind-mount-unix test as it is unsupported in mounts-v2.

https://jira.sw.ru/browse/PSBM-58617

v3: fix nits

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to 3.15 mainstream:
1) reformat according to python standard

2) move the test after ("mount-v2: restore_mount_sharing_options at the
end of forking stage") and other related mount-v2 patches, as now after
60bb5c731 ("zdtm: Set --root path to 0700 on restore") mainstream patch
we can't MS_SET_GROUP from root task as it does not have enough
permissions to open --root path, so moving it to criu task fixes the
problem.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>