mount: use ns_mountpoint in validate_children_collision
Function validate_children_collision is both called on dump and on restore. On
dump mountpoint and ns_mountpoint are the same. On restore as we never call
validate_children_collision on the root_yard_mp, for all other mounts strcmp
results would be the same with mountpoint and ns_mountpoint.

Cherry-picked from Virtuozzo criu:
https://src.openvz.org/projects/OVZ/repos/criu/commits/8f4fda5ac

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>