mount-v2: select wider mount as first in restore_one_sharing_group
This is a fast fix for new MOVE_MOUNT_SET_GROUP restrictions. Now it
also requires source mount root be wider than destination mount root.
So let's atleast try the widest mount we have as a "first" mount.

Actualy there are cases there this won't work - we can have slave mount
with wider root than all mounts of it's master share, we would need to
fix it properly in future.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>