zdtm: add two tests for large ghost sparse file
ghost_holes_large00 is a test which creates a large ghost sparse file
with 10GiB hole and 8KiB data, criu should be able to handle this kind
of situation.

ghost_holes_large01 is a test which creates a large ghost sparse file
with 10GiB hole and 2MiB data, since 2MiB is larger than the default
ghost_limit(1MiB), criu should fail on this test.

Signed-off-by: Liang-Chun Chen <featherclc@gmail.com>