net: dump nft rules wia iptables-nft
We see nft rules with matches like "physdev" or "match-set" which can be
created by iptables-nft but can not be dumped/restored by nft utility.

The idea of this patch is to dump nftables rules both with nft and
iptables-nft. So that we are able to choose which one to use on restore.
The --nftables-mode=ipt/nft option is added to switch between those on
restore (default is ipt).

https://jira.sw.ru/browse/PSBM-138902

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>